Make bt->cache_index an 8-byte quantity as we support database shared memory size more than 4Gib
Background
----------
* This is something that I noticed in the code while reviewing for performance enhancements.

* The `cache_index` field in the `bt_rec` structure is an offset of the corresponding cache-record.

  **sr_port/gdsbt.h**
  ```
  	int4		cache_index;
  ```

* While the `bt_index` field in the `cache_record` structure is an offset to the corresponding
  bt record.

  **sr_port/gdsfhead.h**
  ```
  	sm_off_t	bt_index;       /* offset to corresponding bt_rec ...
  ```

* But notice that both the offsets have differing types. One is a `int4` (a 4-byte offset) whereas
  the other is a `sm_off_t` (an 8-byte offset).

* `sm_off_t` is the right type because this is an offset into a shared memory location and we have
  long supported database shared memory sizes more than 4Gib.

Issue
-----
* So I was wondering if this pointed to a user-visible issue.

  Turns out there is none and it is due to a combination of fortunate unrelated conditions. Below
  is an explanation as to why it is a non-issue.

* One can get a sense of the database shared memory layout by the `dse cache -show` command.
  I created a 9.75Gib database shared memory size with the following settings (lockspace,
  journal buffer size and number of global buffers all set to their maximum allowed values).

  ```
  mupip set -lock=262144 -reg "*"
  mupip set -journal=enable,on,before,buffer_size=1048576 -reg "*"
  mupip set -glo=2097151 -reg "*"
  ```

  This got me the following shared memory layout (`dse cache -show` command output for above database
  is pasted below).

  ```
  DSE> cache -show
  Region DEFAULT : Shared_memory       = 0x000072D8E9906000
  Region DEFAULT :  node_local         = 0x0000000000000000
  Region DEFAULT :  critical           = 0x000000000001F000
  Region DEFAULT :  jnl_buffer_struct  = 0x0000000000020400
  Region DEFAULT :  jnl_buffer_data    = 0x00000000000506E8
  Region DEFAULT :  shmpool_buffer     = 0x0000000020052000
  Region DEFAULT :  lock_space         = 0x0000000020156000
  Region DEFAULT :  cache_queues_state = 0x0000000028156000
  Region DEFAULT :  cache_que_header   = 0x0000000028156030 : Numelems = 0x00200011 : Elemsize = 0x000000E0
  Region DEFAULT :  cache_record       = 0x0000000044156F10 : Numelems = 0x001FFFFF : Elemsize = 0x000000E0
  Region DEFAULT :  global_buffer      = 0x0000000060157000 : Numelems = 0x001FFFFF : Elemsize = 0x00001000
  Region DEFAULT :  db_file_header     = 0x0000000260156000
  Region DEFAULT :  bt_que_header      = 0x0000000260556000 : Numelems = 0x00200011 : Elemsize = 0x00000040
  Region DEFAULT :  th_base            = 0x0000000268556450
  Region DEFAULT :  bt_record          = 0x0000000268556480 : Numelems = 0x001FFFFF : Elemsize = 0x00000040
  Region DEFAULT :  shared_memory_size = 0x0000000270557000
  ```

* Each line corresponds to a section inside shared memory in that order of memory layout.

* Notice that the `cache_record` line comes BEFORE the `global_buffer` line. Whereas the
  `bt_record` line comes AFTER the `global_buffer` line.

* In database shared memory, the section that takes up the most space is the `global_buffer` line.
  That is what can get a lot more than 4Gib.

* The other lines before `global_buffer` can only get so much big as they are currently safeguarded
  from going too high by various unrelated limit checks.

* The `cache_index` value stored in a `bt_rec` is an offset from the start of the `lock_space` line
  to the offset of the Nth cache record in the `cache_record` line.

* The `cache_que_header` line is a sequence of P cache records where P is the first prime number that is
  greater than the user specified number of global buffers for this region.

* The `cache_record` line is a sequence of N cache records where N is the number of user specified
  global buffers for this region.

* So together those 2 lines can take up almost `2 * N` cache records. Each cache-record occupies `224`
  bytes so we need `(2**32)/(224*2)` which is around 9.5 million global buffers for the `cache_index`
  field to become more than 4Gib.

* Currently the maximum number of global buffers that we allow is `2097151` (i.e. `0x1FFFFF`). So the
  current limit is 2 million global buffers. If we ever bump this limit in the future to 10 million
  global buffers, we will end up in trouble due to the `cache_index` offset being a 4-byte quantity.
  The consequences of this can then be memory corruption and even database damage.

* This is a risk best avoided as we will not remember the `cache_index` issue when the time comes
  to allow a higher global buffers limit.

Fix
---
* Therefore this commit makes the `cache_index` field an `8-byte` quantity (i.e. a `sm_off_t` type)
  and that takes care of fixing this futuristic issue.

* This meant that various places in the code where the `cache_index` was being typecast to the
  4-byte `int` or `int4` can now be removed. And that is also taken care of in this commit.

* While at this, I noticed that the adjacent 4-bytes after `cache_index` were a 1-byte `flushing`
  and a 3-byte `filler`. The `flushing` variable is used only in case of cluster support and that
  code has long been unsupported so removed the `flushing` and `filler` variables from the
  `bt_rec` structure and this let me keep the `bt_rec` structure size the same as before inspite
  of the 4-byte increase in the size of the `cache_index` field.

+++ sr_port/t_end_sysops.c
@@ -668,7 +668,7 @@ enum cdb_sc bg_update_phase1(cw_set_element *cs, trans_num ctn)
 		assert(LATCH_SET <= WRITE_LATCH_VAL(cr));
 		BG_TRACE(new_buff);
 		cr->bt_index = GDS_ABS2REL(bt);
-		bt->cache_index = (int4)GDS_ABS2REL(cr);
+		bt->cache_index = GDS_ABS2REL(cr);
 		cr->backup_cr_is_twin = FALSE;
 	} else	/* end of if else on cr NOTVALID */
 	{
